+++ CesiumGltfReader/src/GltfReader.cpp
@@ -596,8 +596,9 @@ void CesiumGltfReader::GltfReader::postprocessGltf(
           } else {
             // We have a depot, so fetch this asset via that depot.
             return options.pSharedAssetSystem->pImage->getOrCreate(
-                asyncSystem,
-                pAssetAccessor,
+                SharedAssetContext{
+                    .asyncSystem = asyncSystem,
+                    .pAssetAccessor = pAssetAccessor},
                 assetKey);
           }
         };
@@ -643,8 +644,9 @@ void CesiumGltfReader::GltfReader::postprocessGltf(
       } else {
         // We have a depot, so fetch this asset via that depot.
         return options.pSharedAssetSystem->pExternalMetadataSchema->getOrCreate(
-            asyncSystem,
-            pAssetAccessor,
+            SharedAssetContext{
+                .asyncSystem = asyncSystem,
+                .pAssetAccessor = pAssetAccessor},
             assetKey);
       }
     };